+++ example/example.dart
@@ -1,5 +1,7 @@
 import 'package:bishop/bishop.dart';
 
+import 'play.dart';
+
 void main(List<String> args) async {
   String v = args.isNotEmpty ? args.first : 'standard';
   Variant variant = variantFromString(v) ?? Variant.standard();
@@ -15,8 +17,9 @@ void main(List<String> args) async {
     print('${Bishop.playerName[game.turn]}: ${game.toSan(m)}');
     game.makeMove(m);
     i++;
-    if (i > 200) break;
+    if (i >= 200) break;
   }
   print(game.ascii());
   print(game.pgn());
+  printCyan(game.result?.readable ?? 'Game Over (too long)');
 }